requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
psycopg2-binary==2.9.9
werkzeug==3.0.3
authlib==1.3.1
requests==2.32.3
reportlab==4.2.2
python-dotenv==1.0.1
pillow==10.4.0
wtforms==3.1.2
email-validator==2.2.0
gunicorn==22.0.0
